+++ dashboard/tests/mocks/database.py
@@ -4,14 +4,14 @@ import sqlite3
 
 class SqliteService:
 
-    def __init__(self):
-        self.conn: sqlite3.Connection = sqlite3.connect(':memory:')
+    def __init__(self, location=':memory:'):
+        self.conn: sqlite3.Connection = sqlite3.connect(location)
         self.conn.row_factory = sqlite3.Row
 
     @staticmethod
     def migrate(client: sqlite3.Connection, migration_script):
         cursor: sqlite3.Cursor = client.cursor()
-        with open(os.path.join(os.path.dirname(os.path.abspath(__file__)), migration_script), 'r') as script:
+        with open(migration_script, 'r') as script:
             cursor.executescript(script.read())
         client.commit()
         cursor.close()