+++ dashboard/utils/__init__.py
@@ -1,6 +1,7 @@
 import re
 import os
 import yaml
+import time
 
 def clean_dag_id(dag_id):
     return re.sub('_v?[0-9]+.[0-9]+$', '', dag_id)
@@ -36,3 +37,27 @@ def get_yaml_file_content(path):
             return yaml.load(file)
     else:
         return None
+
+# source: https://stackoverflow.com/a/1094933/7098262
+def sizeof_fmt(num, suffix='B'):
+    for unit in ['','Ki','Mi','Gi','Ti','Pi','Ei','Zi']:
+        if abs(num) < 1024.0:
+            return "%3.1f%s%s" % (num, unit, suffix)
+        num /= 1024.0
+    return "%.1f%s%s" % (num, 'Yi', suffix)
+
+
+class Timer:
+
+    def __init__(self, logger, description):
+        self.logger = logger
+        self.description = description
+
+    def __enter__(self):
+        self.start = time.time()
+        self.logger.debug(f"{self.description} started")
+        return self
+
+    def __exit__(self, *args):
+        self.interval = time.time() - self.start
+        self.logger.info(f"{self.description} took {self.interval:.2f} sec")